refactor: Rename model package to entity package
