+++ src/main/java/gift/controller/AdminController.java
@@ -1,8 +1,7 @@
 package gift.controller;
 
-import gift.model.ProductRecord;
+import gift.entity.ProductRecord;
 import gift.repository.ProductDAO;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
