+++ src/main/java/gift/controller/ProductController.java
@@ -1,9 +1,8 @@
 package gift.controller;
 
-import gift.model.ProductRecord;
+import gift.entity.ProductRecord;
 import gift.repository.ProductDAO;
 import jakarta.validation.Valid;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
