+++ src/main/java/gift/entity/User.java
@@ -1,0 +1,7 @@
+package gift.entity;
+
+public class User {
+    private int id;
+    private String email;
+    private String password;
+}
